Collect RLimit for only linux and darwin
