+++ internal/statistics_osmetric.go
@@ -1,0 +1,31 @@
+// Copyright (c) 2008-2018, Hazelcast, Inc. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License")
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// +build linux darwin
+
+package internal
+
+import (
+	"bytes"
+	"strconv"
+	"syscall"
+)
+
+func (s *statistics) collectOSMetrics(stats *bytes.Buffer) {
+	var limit syscall.Rlimit
+	if err := syscall.Getrlimit(syscall.RLIMIT_NOFILE, &limit); err == nil {
+		s.addStat(stats, "os.maxFileDescriptorCount", strconv.Itoa(int(limit.Max)))
+		s.addStat(stats, "os.openFileDescriptorCount", strconv.Itoa(int(limit.Cur)))
+	}
+}
